fixed most if not all bugs that i found, moved around the code in lawmaking
i just kept fixing things and running through the code with 9998 iterations (plus gen_1() and gen_2() of course) until it worked. the program now works and i believe it will be the simplest version of the simulation that is still an accurate description of the system. comments are all fixed and explain the important parts of the code.